Conflicts:
	app/database/migrations/2014_02_10_080351_confide_setup_users_table.php
	app/database/migrations/2014_02_21_131118_product.php
	app/routes.php

+++ app/controllers/ProductController.php
@@ -2,6 +2,7 @@
 
 class ProductController extends BaseController {
 
+
 	/**
 	 * Display a listing of the resource.
 	 *
@@ -41,14 +42,14 @@ class ProductController extends BaseController {
 		$rules = array(
 			'name'       => 'required',
 
-		);
+			);
 		$validator = Validator::make(Input::all(), $rules);
 
 		// process the login
 		if ($validator->fails()) {
 			return Redirect::to('product/create')
-				->withErrors($validator)
-				->withInput();
+			->withErrors($validator)
+			->withInput();
 		} else {
 			// store
 			$product = new prod;
@@ -56,16 +57,40 @@ class ProductController extends BaseController {
 			$product->price         = Input::get('price');
 			$product->brand_id      = Input::get('brand');
 			$product->category_id   = Input::get('category');
-			$image = Input::file('product_pic');
-			$filename = date('Y-m-d-H-i-s')."-".$image->getClientOriginalName();
-			Image::make($image->getRealPath())->save(public_path().'/img/products/'.$filename);
 
-			$product->product_pic = $filename;
+
+
+			if(Input::hasFile('product_pic')){
+
+				$image = Input::file('product_pic');
+				$filename = date('Y-m-d-H-i-s')."-".$image->getClientOriginalName();
+				Image::make($image->getRealPath())->save(public_path().'/img/products/'.$filename);
+				$product->product_pic = 'img/products/'.$filename;
+
+			}else if (Input::has('product_pic')) {
+				$product->product_pic = Input::get('product_pic');
+			}
+
 			$product->save();
-			var_dump($product);
-			// redirect
+
+			//waiting for edit
+
+			
+			for ($i=0; Input::has('type_'.$i); $i++) { 
+				for ($j=0; Input::has('att_'.$i.$j); $j++) { 
+					$att = new Attribute;
+					$att->name = Input::get('att_'.$i.$j);
+					$att->type = Input::get('type_'.$i);
+					$att->product_id = $product->id;
+					$att->save();
+				}
+			}
+
+
+
+
 			Session::flash('message', 'Successfully created product!');
-			return Redirect::to('product');
+			return Redirect::to('product/create');
 		}
 
 	}
@@ -77,11 +102,30 @@ class ProductController extends BaseController {
 	 * @return Response
 	 */
 	public function show($id)
-	{
+	{	Attribute::unguard();
 		$product = Prod::find($id);
+		$temp = Attribute::where('product_id', '=', $id)
+		->get()
+		->toJson();
+
+		$temp = json_decode($temp);
+		$atts = array();
+		foreach ($temp as $value) {
+			$index = count($atts);
+			for ($i=0; $i <count($atts) ; $i++) { 
+				if($value->type == $atts[$i]['name']) {
+					$index = $i;
+					break;
+				}
+			}
+
+			if($index == count($atts)) $atts[] = array('name' => $value->type, 'data' => array( $value->name) );
+			else $atts[$index]['data'][] = $value->name;
+
+		}
 
 		return View::make('pages.product.show')
-			->with('product', $product);
+		->with(array('product'=> $product, 'atts'=>$atts ));
 
 	}
 
@@ -98,7 +142,29 @@ class ProductController extends BaseController {
 		$brand_all = Brand::All();
 		$category_all = Category::All();
 
-		return View::make('pages.product.edit',array( 'product'=> $product, 'brand_all' => $brand_all, 'category_all' => $category_all ) );
+
+		$temp = Attribute::where('product_id', '=', $id)
+		->get()
+		->toJson();
+
+		$temp = json_decode($temp);
+		$atts = array();
+		foreach ($temp as $value) {
+			$index = count($atts);
+			for ($i=0; $i <count($atts) ; $i++) { 
+				if($value->type == $atts[$i]['name']) {
+					$index = $i;
+					break;
+				}
+			}
+
+			if($index == count($atts)) $atts[] = array('name' => $value->type, 'data' => array( $value->name) );
+			else $atts[$index]['data'][] = $value->name;
+
+		}
+
+		// var_dump(json_encode($atts));
+		return View::make('pages.product.edit',array( 'product'=> $product, 'brand_all' => $brand_all, 'category_all' => $category_all, 'atts' => $atts ) );
 
 	}
 
@@ -113,14 +179,14 @@ class ProductController extends BaseController {
 		$rules = array(
 			'name'       => 'required',
 
-		);
+			);
 		$validator = Validator::make(Input::all(), $rules);
 
 		// process the login
 		if ($validator->fails()) {
 			return Redirect::to('product/create')
-				->withErrors($validator)
-				->withInput();
+			->withErrors($validator)
+			->withInput();
 		} else {
 			// store
 			$product = Prod::find($id);
@@ -128,13 +194,39 @@ class ProductController extends BaseController {
 			$product->price         = Input::get('price');
 			$product->brand_id      = Input::get('brand');
 			$product->category_id   = Input::get('category');
-			File::delete(public_path().$product->image);
+			
+			
+			
+			if(Input::hasFile('product_pic')){
 
-			$image = Input::file('product_pic');
-			$filename = date('Y-m-d-H-i-s')."-".$image->getClientOriginalName();
-			Image::make($image->getRealPath())->save(public_path().'/img/products/'.$filename);
-			$product->product_pic = $filename;
+				$image = Input::file('product_pic');
+				$filename = date('Y-m-d-H-i-s')."-".$image->getClientOriginalName();
+				Image::make($image->getRealPath())->save(public_path().'/img/products/'.$filename);
+				$product->product_pic = 'img/products/'.$filename;
+
+			}else if (Input::has('product_pic')) {
+				if(File::exists(public_path().$product->product_pic))
+					File::delete(public_path().$product->image);
+				$product->product_pic = Input::get('product_pic');
+			}
+
+
 			$product->save();
+
+
+			for ($i=0; Input::has('type_'.$i); $i++) { 
+				for ($j=0; Input::has('att_'.$i.$j); $j++) { 
+					$att = new Attribute;
+					$att->name = Input::get('att_'.$i.$j);
+					$att->type = Input::get('type_'.$i);
+					$att->product_id = $product->id;
+					if(!Attribute::where('name','=',$att->name)
+						->where('type','=',$att->type)
+						->where('product_id','=',$att->product_id))
+						$att->save();
+				}
+			}
+
 
 			// redirect
 			Session::flash('message', 'Successfully update product!');
@@ -155,7 +247,7 @@ class ProductController extends BaseController {
 			File::delete(public_path().$product->image);
 			$product->delete();
 			Session::flash('message', 'Successfully deleted the product!');
-		return Redirect::to('product');
+			return Redirect::to('product');
 		}
